Conflicts:
	pom-zk33.xml
	pom-zk34.xml
	src/main/java/com/twitter/distributedlog/AppendOnlyStreamWriter.java
	src/main/java/com/twitter/distributedlog/BKBaseLogWriter.java
	src/main/java/com/twitter/distributedlog/BKContinuousLogReaderBase.java
	src/main/java/com/twitter/distributedlog/BKDistributedLogManager.java
	src/main/java/com/twitter/distributedlog/BKLogPartitionHandler.java
	src/main/java/com/twitter/distributedlog/BKLogPartitionReadHandler.java
	src/main/java/com/twitter/distributedlog/BKLogPartitionWriteHandler.java
	src/main/java/com/twitter/distributedlog/BKPartitionAwareLogWriter.java
	src/main/java/com/twitter/distributedlog/BKPerStreamLogReader.java
	src/main/java/com/twitter/distributedlog/BKPerStreamLogWriter.java
	src/main/java/com/twitter/distributedlog/BKUnPartitionedLogWriterBase.java
	src/main/java/com/twitter/distributedlog/BKUnPartitionedSyncLogWriter.java
	src/main/java/com/twitter/distributedlog/DistributedLogManager.java
	src/main/java/com/twitter/distributedlog/LedgerDataAccessor.java
	src/main/java/com/twitter/distributedlog/LogReader.java
	src/main/java/com/twitter/distributedlog/LogRecord.java
	src/main/java/com/twitter/distributedlog/LogSegmentLedgerMetadata.java
	src/main/java/com/twitter/distributedlog/PerStreamLogReader.java
	src/main/java/com/twitter/distributedlog/ResumableBKPerStreamLogReader.java
	src/test/java/com/twitter/distributedlog/TestBookKeeperDistributedLogManager.java
	src/test/java/com/twitter/distributedlog/TestInterleavedReaders.java

+++ src/main/java/com/twitter/distributedlog/PartitionId.java
@@ -22,7 +22,7 @@ public class PartitionId {
 
     @Override
     public int hashCode() {
-        return (int) (partition * 13 ^ partition * 17);
+        return (partition * 13 ^ partition * 17);
     }
 
     @Override
